+++ packages/cli/src/deploy/cloud-functions/getFunctionsData.ts
@@ -3,11 +3,13 @@ import { writeFileSync } from "fs"
 import { generateSchema } from "../schema/generateSchema.js"
 import { join } from "path"
 import { GenerateGlobalConfig } from "@stoker-platform/types"
+import { pathToFileURL } from "url"
 
 export const getFunctionsData = async () => {
     await initializeFirebase()
-    const configFilePath = join(process.cwd(), "lib", "main.js")
-    const globalConfigFile = await import(/* @vite-ignore */ configFilePath)
+    const path = join(process.cwd(), "lib", "main.js")
+    const url = pathToFileURL(path).href
+    const globalConfigFile = await import(/* @vite-ignore */ url)
     const config: GenerateGlobalConfig = globalConfigFile.default
     const globalConfig = config("node")
     const timezone = await tryPromise(globalConfig.timezone)